+++ js/draw-big-picture.js
@@ -1,10 +1,72 @@
 import { isEscapeKey } from './util.js';
-import { renderComments } from './render-data.js';
-import { onLoadBtnClick, renderMainData, loadBtn } from './render-data.js';
+
+const STEP_COMMENTS = 5;
 
 const bigPicture = document.querySelector('.big-picture');
 const closeBtn = document.querySelector('.big-picture__cancel');
 const body = document.querySelector('body');
+const thumbnailsContainerElement = document.querySelector('.social__comments');
+const loadComment = document.querySelector('.social__comment-count');
+const loadCommentCount = document.querySelector('.js-load-comment-count');
+const loadBtn = document.querySelector('.comments-loader');
+
+let currentComments = null;
+let sliceCurrentComments = null;
+let endIndex = 0;
+
+const bigPictureInfo = {
+  img: document.querySelector('.big-picture__img').querySelector('img'),
+  likes: document.querySelector('.likes-count'),
+  comments: document.querySelector('.comments-count'),
+  description: document.querySelector('.social__caption')
+};
+
+const getCommentTemplate = (comment) => `<li class="social__comment">
+<img class="social__picture" src=${comment.avatar} alt=${comment.name} width="35" height="35">
+<p class="social__text">${comment.message}</p>
+</li>`;
+
+const thumbnailsInit = () => {
+  endIndex = Math.min(currentComments.length, endIndex + STEP_COMMENTS);
+  sliceCurrentComments = currentComments.slice(0, endIndex);
+
+  thumbnailsContainerElement.innerHTML = '';
+
+  if (sliceCurrentComments) {
+    thumbnailsContainerElement.insertAdjacentHTML('beforeend', sliceCurrentComments.map((element) => getCommentTemplate(element)).join(''));
+    loadCommentCount.textContent = `${endIndex} из ${currentComments.length} комментариев`;
+  }
+
+  if (endIndex === currentComments.length) {
+    loadBtn.classList.add('hidden');
+  }
+};
+
+const renderMainData = (photo) => {
+  bigPictureInfo.img.src = photo.url;
+  bigPictureInfo.likes.textContent = photo.likes;
+  bigPictureInfo.comment = photo.comments.length;
+  bigPictureInfo.description.textContent = photo.description;
+};
+
+const onLoadBtnClick = () => thumbnailsInit(currentComments);
+
+const renderComments = (comments) => {
+  endIndex = 0;
+  currentComments = comments;
+
+  thumbnailsInit(comments);
+
+  if (comments.length <= STEP_COMMENTS) {
+    loadBtn.classList.add('hidden');
+    loadComment.classList.add('hidden');
+  } else {
+    loadBtn.classList.remove('hidden');
+    loadComment.classList.remove('hidden');
+
+    loadBtn.addEventListener('click', onLoadBtnClick);
+  }
+};
 
 const onDocumentKeydown = (evt) => {
   if (isEscapeKey(evt)) {
@@ -23,7 +85,7 @@ function closeViewPopup() {
   loadBtn.removeEventListener('click', onLoadBtnClick);
 }
 
-export const openPopup = (photo) => {
+export const openViewPopup = (photo) => {
   renderMainData(photo);
   renderComments(photo.comments);
