+++ js/draw-pictures.js
@@ -1,16 +1,8 @@
-import { openPopup } from './draw-big-picture.js';
+import { openViewPopup } from './draw-big-picture.js';
 
 const template = document.querySelector('#picture').content.querySelector('.picture');
 const fragment = document.createDocumentFragment();
 const picturesContainer = document.querySelector('.pictures');
-
-const openViewPopup = (photos) => {
-  const pictures = document.querySelectorAll('.picture');
-
-  pictures.forEach((picture, index) => {
-    picture.addEventListener('click', () => openPopup(photos[index]));
-  });
-};
 
 export const drawPictures = (photos) => {
   photos.forEach((photo) => {
@@ -29,6 +21,9 @@ export const drawPictures = (photos) => {
   });
 
   picturesContainer.appendChild(fragment);
+  const pictures = document.querySelectorAll('.picture');
 
-  openViewPopup(photos);
+  pictures.forEach((picture, index) => {
+    picture.addEventListener('click', () => openViewPopup(photos[index]));
+  });
 };
