+++ js/main.js
@@ -1,4 +1,4 @@
-import { drawPictures } from './draw.js';
+import { drawPictures } from './draw-pictures.js';
 import { generateDataForPhoto } from './data.js';
 
 const COUNT_PHOTOS = 25;
